## Description

WEB-802

There is a feature to search workspaces and projects in Web UI. [`GetWorkspaceProjects`](https://github.com/determined-ai/determined/blob/990e62e958d8bf29ac2e58ffb49d2be62ab15a90/master/internal/api_workspace.go#L131) API is called to fetch data. However, it was critically slow to fetch data when there are many workspaces and projects, and returned dupe data.

- Before the fix, it takes around **5 min** to get projects associated with all workspaces
- After the fix, it only takes less than **1 sec**.

I replaced `subquery` with `JOIN` to speed up.

As a result 

|                            | Old Planning Time | New Planning Time | Speed Up Rate |
| --------------- | -------------------: | --------------------: | --------------:  |
| All workspaces | 0.820 ms                  | 7.629 ms                     | -89.251% 📉     |
| A workspace     | 1.063 ms                  | 0.880 ms                    | 120.795% 📈     |

|                            | Old Execution Time | New Execution Time | Speed Up Rate    |
| --------------- | -------------------: | --------------------: | ---------------:  |
| All workspaces | 321442.007 ms        | 2.266 ms                    | 14,185,437% 📈  |
| A workspace     | 105.973 ms               | 0.342 ms                    | 30,986% 📈        |


Overall = Planning Time + Execution Time

|                            | Old Overall Time      | New Overall Time      | Speed Up Rate |
| --------------- | -------------------: | --------------------: | --------------:  |
| All workspaces | 321442.827 ms        | 9.895 ms                    | 3,248,537% 📈 |
| A workspace     | 107.036 ms               | 1.222 ms                    | 8,759% 📈         |

For more details of benchmarks, see [**the comment below**](https://github.com/determined-ai/determined/pull/5744#issuecomment-1382693174)

## Screenshots

### Before

https://user-images.githubusercontent.com/109113616/212441124-21457f0b-7ca0-4a44-a3d9-b8ca764c22c2.mov

### After

https://user-images.githubusercontent.com/109113616/212441004-ee5e2ea6-d156-448c-a27c-3871d4433da8.mov



<!---
Lead with the intended commit body in this description field. For breaking
changes, please include "BREAKING CHANGE:" at the beginning of your commit
body.  At a minimum, this section should include a bracketed reference to the
Jira ticket, e.g. "[DET-1234]". When squash-and-merging, copy this directly
into the description field.
-->



## Test Plan
- Check if API returns the same data as before with the same params  
  - `id=0` means getting all data
  - `id= > 1` means getting a specific workspace id data
- Check if query is fast enough
- Check if data are not duplicated
- Check if it works as before with RBAC enabled
<!---
Describe the situations in which you've tested your change, and/or a screenshot
as appropriate.  Reviewers may ask questions about this test plan to ensure
adequate manual coverage of changes.
-->



## Commentary (optional)

<!---
Use this section of your description to add context to the PR. Could be for
particularly tricky bits of code that could use extra scrutiny, historical
context useful for reviewers, etc.
You may intentionally leave this section blank and remove the title.
--->



## Checklist

- [x] Changes have been manually QA'd
- [x] User-facing API changes need the "User-facing API Change" label.
- [x] Release notes should be added as a separate file under `docs/release-notes/`.
  See [Release Note](https://github.com/determined-ai/determined/blob/master/docs/release-notes/README.md) for details.
- [x] Licenses should be included for new code which was copied and/or modified from any external code.

## Ticket
<!---
Retain the relevant line and replace 000 with ticket number.

DET-000
MLG-000
WEB-000
DESIGN-000
No Ticket
--->


<!---
## Title

Example title: "docs: tweak recommended "pip install" usage".

Specifically, this title should contain a type and a description
of the change being made:

User-facing change types:

- docs: docs-only change
- feat: new user-facing feature
- fix: bug fix
- perf: performance improvement

Internal change types:

- build: build system change (anything in a `Makefile`, mostly)
- chore: any internal change not covered by another type
- ci: anything that touches `.circleci`
- refactor: internal refactor
- style: style change
- test: new tests

See https://www.conventionalcommits.org/en/v1.0.0/ for background.

The first line should also:

- be at most 89 characters long
- contain a description that is at most 72 characters long
- not end with sentence-ending punctuation
- start (after the type) with a lowercase imperative ("add", "fix")

-->